// client/src/api/models/postDetailsModel.ts
/**
 * Generated by orval v6.10.3 🍺
 * Do not edit manually.
 * Readdit
 * OpenAPI spec version: v1
 */
import type { FeedPostAuthorModel } from "./feedPostAuthorModel";
import type { FeedCommunityModel } from "./feedCommunityModel";
import type { UserVoteModel } from "./userVoteModel";
import type { PostCommentDetailsModel } from "./postCommentDetailsModel";

export interface PostDetailsModel {
   id?: string | null;
   author?: FeedPostAuthorModel;
   community?: FeedCommunityModel;
   mediaUrl?: string | null;
   mediaPublicId?: string | null;
   title?: string | null;
   content?: string | null;
   voteScore?: number;
   createdOn?: string;
   modifiedOn?: string | null;
   userVote?: UserVoteModel;
   comments?: PostCommentDetailsModel[] | null;
   tags?: string[] | null;
}