// client/src/api/models/postVote.ts
/**
 * Generated by orval v6.10.3 🍺
 * Do not edit manually.
 * Readdit
 * OpenAPI spec version: v1
 */
import type { CommunityPost } from "./communityPost";
import type { ApplicationUser } from "./applicationUser";
import type { VoteType } from "./voteType";

export interface PostVote {
   id?: string | null;
   createdOn?: string;
   modifiedOn?: string | null;
   isDeleted?: boolean;
   deletedOn?: string | null;
   postId: string;
   post?: CommunityPost;
   userId: string;
   user?: ApplicationUser;
   type: VoteType;
}